Here are changes to the project prepare to revert. Just in case. Prepare to revert.
I changed the SnowLevel AI Placement. I added the possibility to damage CaptainPenguin with the projectile. I added the possibility to deal projectile damage to the FlyingAI. I changed the  hud's mana regeneration. I allowed the
enemy's fireball to be destroyed on impact with the main character. I added more AI and medpacks to the LavaLevel.